+++ src/main/java/com/simpleweb/model/Post.java
@@ -14,8 +14,8 @@ import java.util.Date;
 public class Post {
     public enum Status { ACTIVE, INACTIVE };
 
-    private int id;
-    private int userId;
+    private Long id;
+    private Long userId;
     private String title;
     private String details;
     private Status status;
@@ -24,22 +24,23 @@ public class Post {
     private Date created;
     private Date updated;
 
+
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id", unique = true)
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
